openenv/requirements.txt
pydantic>=2.10.0
openai>=1.54.0
python-dotenv>=1.0.1
